requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
html2text==2018.1.9
idna==2.7
itsdangerous==0.24
Jinja2==2.10
ldap3==2.5.1
html5lib==1.0.1
lxml==4.2.4
MarkupSafe==1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-ldap==3.1.0
requests==2.19.1
robobrowser==0.5.3
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
PyPDF2==1.26.0
opencv-python==3.4.3.18
camelot-py==0.2.3
PyJWT==1.6.4
bcrypt==3.1.4
